break down gem_sui/lib to smaller files
